Update ElasticTransformation and add ElasticTransform
