+++ src/ui/dtl/__init__.py
@@ -112,7 +112,13 @@ from .actions.output.create_labeling_job.create_labeling_job import CreateLabeli
 from .actions.input.filtered_project.filtered_project import FilteredProjectAction
 from .actions.other.move.move import MoveAction
 from .actions.other_augs.pixelate.pixelate import PixelateAction
-from .actions.other_augs.elastic_transform.elastic_transform import ElasticTransformAction
+from .actions.other_augs.elastic_transformation.elastic_transformation import (
+    ElasticTransformationAction,
+)
+
+from .actions.other_augs.elastic_transform.elastic_transform import (
+    ElasticTransformAction,
+)
 
 import src.globals as g
 
@@ -152,7 +158,11 @@ image_actions_list = {
         RotateAction.name,
         SlidingWindowAction.name,
     ],
-    OTHER_AUGMENTATIONS: [PixelateAction.name, ElasticTransformAction.name],
+    OTHER_AUGMENTATIONS: [
+        PixelateAction.name,
+        ElasticTransformAction.name,
+        ElasticTransformationAction.name,
+    ],
     ANNOTATION_TRANSFORMS: [
         ApproxVectorAction.name,
         BackgroundAction.name,
@@ -217,6 +227,7 @@ image_actions_dict = {
     # Other Augmentations
     PixelateAction.name: PixelateAction,
     ElasticTransformAction.name: ElasticTransformAction,
+    ElasticTransformationAction.name: ElasticTransformationAction,
     # Annotation layers
     ApproxVectorAction.name: ApproxVectorAction,
     BackgroundAction.name: BackgroundAction,
